refactored for better pattern
