--- apps/website/components/templates/landing/product-show/styles.tsx
@@ -1,8 +1,0 @@
-import Image from 'next/image';
-
-import styled from '@emotion/styled';
-
-export const ResponsiveImage = styled(Image)`
-  max-width: 100%;
-  height: auto;
-`;
